lab7 ver2
