+++ labs/lab7/src/oil-prices/brent-year.json
@@ -1,0 +1,146 @@
+[
+  {
+    "Date": "1987-06-30",
+    "Price": "18.53"
+  },
+  {
+    "Date": "1988-06-30",
+    "Price": "14.91"
+  },
+  {
+    "Date": "1989-06-30",
+    "Price": "18.23"
+  },
+  {
+    "Date": "1990-06-30",
+    "Price": "23.76"
+  },
+  {
+    "Date": "1991-06-30",
+    "Price": "20.04"
+  },
+  {
+    "Date": "1992-06-30",
+    "Price": "19.32"
+  },
+  {
+    "Date": "1993-06-30",
+    "Price": "17.01"
+  },
+  {
+    "Date": "1994-06-30",
+    "Price": "15.86"
+  },
+  {
+    "Date": "1995-06-30",
+    "Price": "17.02"
+  },
+  {
+    "Date": "1996-06-30",
+    "Price": "20.64"
+  },
+  {
+    "Date": "1997-06-30",
+    "Price": "19.11"
+  },
+  {
+    "Date": "1998-06-30",
+    "Price": "12.76"
+  },
+  {
+    "Date": "1999-06-30",
+    "Price": "17.9"
+  },
+  {
+    "Date": "2000-06-30",
+    "Price": "28.66"
+  },
+  {
+    "Date": "2001-06-30",
+    "Price": "24.46"
+  },
+  {
+    "Date": "2002-06-30",
+    "Price": "24.99"
+  },
+  {
+    "Date": "2003-06-30",
+    "Price": "28.85"
+  },
+  {
+    "Date": "2004-06-30",
+    "Price": "38.26"
+  },
+  {
+    "Date": "2005-06-30",
+    "Price": "54.57"
+  },
+  {
+    "Date": "2006-06-30",
+    "Price": "65.16"
+  },
+  {
+    "Date": "2007-06-30",
+    "Price": "72.44"
+  },
+  {
+    "Date": "2008-06-30",
+    "Price": "96.94"
+  },
+  {
+    "Date": "2009-06-30",
+    "Price": "61.74"
+  },
+  {
+    "Date": "2010-06-30",
+    "Price": "79.61"
+  },
+  {
+    "Date": "2011-06-30",
+    "Price": "111.26"
+  },
+  {
+    "Date": "2012-06-30",
+    "Price": "111.63"
+  },
+  {
+    "Date": "2013-06-30",
+    "Price": "108.56"
+  },
+  {
+    "Date": "2014-06-30",
+    "Price": "98.97"
+  },
+  {
+    "Date": "2015-06-30",
+    "Price": "52.32"
+  },
+  {
+    "Date": "2016-06-30",
+    "Price": "43.64"
+  },
+  {
+    "Date": "2017-06-30",
+    "Price": "54.13"
+  },
+  {
+    "Date": "2018-06-30",
+    "Price": "71.34"
+  },
+  {
+    "Date": "2019-06-30",
+    "Price": "64.3"
+  },
+  {
+    "Date": "2020-06-30",
+    "Price": "41.96"
+  },
+  {
+    "Date": "2021-06-30",
+    "Price": "70.86"
+  },
+  {
+    "Date": "2022-06-30",
+    "Price": "100.93"
+  }
+]
