+++ labs/lab7/src/oil-prices/wti-year.json
@@ -1,0 +1,150 @@
+[
+  {
+    "Date": "1986-06-30",
+    "Price": "15.05"
+  },
+  {
+    "Date": "1987-06-30",
+    "Price": "19.2"
+  },
+  {
+    "Date": "1988-06-30",
+    "Price": "15.97"
+  },
+  {
+    "Date": "1989-06-30",
+    "Price": "19.64"
+  },
+  {
+    "Date": "1990-06-30",
+    "Price": "24.53"
+  },
+  {
+    "Date": "1991-06-30",
+    "Price": "21.54"
+  },
+  {
+    "Date": "1992-06-30",
+    "Price": "20.58"
+  },
+  {
+    "Date": "1993-06-30",
+    "Price": "18.43"
+  },
+  {
+    "Date": "1994-06-30",
+    "Price": "17.2"
+  },
+  {
+    "Date": "1995-06-30",
+    "Price": "18.43"
+  },
+  {
+    "Date": "1996-06-30",
+    "Price": "22.12"
+  },
+  {
+    "Date": "1997-06-30",
+    "Price": "20.61"
+  },
+  {
+    "Date": "1998-06-30",
+    "Price": "14.42"
+  },
+  {
+    "Date": "1999-06-30",
+    "Price": "19.34"
+  },
+  {
+    "Date": "2000-06-30",
+    "Price": "30.38"
+  },
+  {
+    "Date": "2001-06-30",
+    "Price": "25.98"
+  },
+  {
+    "Date": "2002-06-30",
+    "Price": "26.18"
+  },
+  {
+    "Date": "2003-06-30",
+    "Price": "31.08"
+  },
+  {
+    "Date": "2004-06-30",
+    "Price": "41.51"
+  },
+  {
+    "Date": "2005-06-30",
+    "Price": "56.64"
+  },
+  {
+    "Date": "2006-06-30",
+    "Price": "66.05"
+  },
+  {
+    "Date": "2007-06-30",
+    "Price": "72.34"
+  },
+  {
+    "Date": "2008-06-30",
+    "Price": "99.67"
+  },
+  {
+    "Date": "2009-06-30",
+    "Price": "61.95"
+  },
+  {
+    "Date": "2010-06-30",
+    "Price": "79.48"
+  },
+  {
+    "Date": "2011-06-30",
+    "Price": "94.88"
+  },
+  {
+    "Date": "2012-06-30",
+    "Price": "94.05"
+  },
+  {
+    "Date": "2013-06-30",
+    "Price": "97.98"
+  },
+  {
+    "Date": "2014-06-30",
+    "Price": "93.17"
+  },
+  {
+    "Date": "2015-06-30",
+    "Price": "48.66"
+  },
+  {
+    "Date": "2016-06-30",
+    "Price": "43.29"
+  },
+  {
+    "Date": "2017-06-30",
+    "Price": "50.8"
+  },
+  {
+    "Date": "2018-06-30",
+    "Price": "65.23"
+  },
+  {
+    "Date": "2019-06-30",
+    "Price": "56.99"
+  },
+  {
+    "Date": "2020-06-30",
+    "Price": "39.16"
+  },
+  {
+    "Date": "2021-06-30",
+    "Price": "68.13"
+  },
+  {
+    "Date": "2022-06-30",
+    "Price": "94.9"
+  }
+]
